seaprating login and register routes in auth
