+++ src/models/user.js
@@ -1,5 +1,4 @@
 const jwt = require("jsonwebtoken");
-// const config = require("config");
 const config = require("dotenv").config();
 
 const { Model } = require("sequelize");
